requirements.txt
alembic==1.13.2
annotated-types==0.7.0
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
Faker==28.0.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy==1.11.1
mypy-extensions==1.0.0
mysqlclient==2.2.4
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
six==1.16.0
SQLAlchemy==2.0.32
starkbank-ecdsa==2.2.0
tomli==2.0.1
typing_extensions==4.12.2
waitress==3.0.0
Werkzeug==3.0.3
zipp==3.20.1
